+++ domain/src/main/java/com/team23/achievements/repositories/FactRepository.kt
@@ -1,4 +1,4 @@
-package com.team23.achievements.domain.repositories
+package com.team23.achievements.repositories
 
 interface FactRepository {
     suspend fun countReadFacts(): Int